Split apart withdrawals flow tests. Add gatekeep tests
