+++ src/components/flows/withdrawals/test/WithdrawalsMainFlow.test.tsx
@@ -1,17 +1,29 @@
 import { setupServer } from 'msw/node';
-import { screen, waitForElementToBeRemoved, within } from '@testing-library/react';
+import { screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { Route } from 'react-router-dom';
 import { createMemoryHistory, History } from 'history';
-import { initializeConfiguration } from '../../config/config';
-import LoggedInApp from '../../LoggedInApp';
-import { createDefaultStore, login, renderWrapped } from '../../../test/utils';
+import { initializeConfiguration } from '../../../config/config';
+import LoggedInApp from '../../../LoggedInApp';
+import { createDefaultStore, login, renderWrapped } from '../../../../test/utils';
 import {
   pensionAccountStatementBackend,
   useTestBackendsExcept,
   withdrawalsEligibilityBackend,
-} from '../../../test/backend';
-import { WithdrawalMandateDetails } from '../../common/apiModels/withdrawals';
+} from '../../../../test/backend';
+import {
+  assertDoneScreenFundPension,
+  assertDoneScreenPartialWithdrawal,
+  assertDoneScreenSecondPillarWarning,
+  assertFundPensionCalculations,
+  assertFundPensionMandate,
+  assertMandateCount,
+  assertPartialWithdrawalMandate,
+  assertTotalTaxText,
+  confirmAndSignAndAssertDone,
+  enterIban,
+  nextButton,
+} from './utils';
 
 const server = setupServer();
 let history: History;
@@ -497,429 +509,3 @@ describe('withdrawals flow with only third pillar', () => {
     await confirmAndSignAndAssertDone();
   });
 });
-
-describe('withdrawals flow before early retirement age', () => {
-  beforeEach(() => {
-    pensionAccountStatementBackend(server);
-    withdrawalsEligibilityBackend(server, {
-      age: 25,
-      hasReachedEarlyRetirementAge: false,
-      canWithdrawThirdPillarWithReducedTax: false,
-      recommendedDurationYears: 55,
-      arrestsOrBankruptciesPresent: true,
-    });
-  });
-
-  test('can only make calculations on first step and not proceed', async () => {
-    expect(await screen.findByText(/25 years old/i)).toBeInTheDocument();
-    expect(await screen.findByText(/in 35 years/i)).toBeInTheDocument();
-    expect(await screen.findByText(/you might receive/i)).toBeInTheDocument();
-
-    expect(
-      await screen.findByText(/Withdraw from the entire pension holding/i),
-    ).toBeInTheDocument();
-
-    await assertFundPensionCalculations(
-      '457.19 € per month',
-      '0.38%',
-      /will earn returns for the next 22 years/i,
-    );
-
-    const partialWithdrawalSizeInput = await screen.findByLabelText(
-      'Do you wish to make a partial withdrawal immediately?',
-      { exact: false },
-    );
-
-    userEvent.type(partialWithdrawalSizeInput, '20000');
-    assertTotalTaxText('−2 000.00 €');
-
-    await assertFundPensionCalculations(
-      '381.44 € per month',
-      '0.38%',
-      /will earn returns for the next 22 years/i,
-    );
-
-    expect(nextButton()).toBeDisabled();
-
-    userEvent.click(nextButton());
-
-    expect(
-      await screen.findByText(/Withdraw from the entire pension holding/i),
-    ).toBeInTheDocument();
-  });
-});
-
-describe('withdrawals flow at 55 withdrawing only third pillar', () => {
-  beforeEach(() => {
-    pensionAccountStatementBackend(server);
-    withdrawalsEligibilityBackend(server, {
-      age: 55,
-      hasReachedEarlyRetirementAge: false,
-      canWithdrawThirdPillarWithReducedTax: true,
-      recommendedDurationYears: 5 + 20, // 5 years to go until 60 + 20 years recommended duration
-      arrestsOrBankruptciesPresent: false,
-    });
-  });
-
-  test('can only withdraw third pillar', async () => {
-    expect(await screen.findByText(/55 years old/i)).toBeInTheDocument();
-
-    expect(
-      await screen.findByText(/eligible to start using your III pillar holdings/i),
-    ).toBeInTheDocument();
-
-    expect(
-      await screen.findByText(
-        /You will be able to use your II pillar holdings under preferential conditions when you are 60 years old./i,
-        undefined,
-        {
-          timeout: 1000,
-        },
-      ),
-    ).toBeInTheDocument();
-
-    userEvent.click(await screen.findByLabelText(/Withdraw only from II pillar/i));
-
-    expect(
-      await screen.findByText(/Withdraw only from III pillar/i, undefined, {
-        timeout: 1000,
-      }),
-    ).toBeInTheDocument();
-
-    await assertFundPensionCalculations(
-      '19.00 € per month',
-      '0.33%',
-      /will earn returns for the next 25 years/i,
-    );
-
-    const partialWithdrawalSizeInput = await screen.findByLabelText(
-      'Do you wish to make a partial withdrawal immediately?',
-      { exact: false },
-    );
-
-    userEvent.type(partialWithdrawalSizeInput, '1000');
-    assertTotalTaxText('−100.00 €');
-
-    await assertFundPensionCalculations(
-      '15.66 € per month',
-      '0.33%',
-      /will earn returns for the next 25 years/i,
-    );
-
-    userEvent.click(nextButton());
-
-    await enterIban('EE591254471322749514');
-    userEvent.click(nextButton());
-
-    expect(
-      await screen.findByText(/I submit the following applications and am aware of their terms/i),
-    ).toBeInTheDocument();
-
-    expect(await screen.findByText(/EE591254471322749514/i)).toBeInTheDocument();
-    expect(await screen.findByText('EST')).toBeInTheDocument();
-
-    assertMandateCount(2);
-
-    await assertFundPensionMandate('THIRD', '15.66 €');
-    await assertPartialWithdrawalMandate(
-      'THIRD',
-      [
-        {
-          fundName: 'Tuleva III Samba Pensionifond',
-          liquidationAmount: '1279.92 units',
-        },
-      ],
-      '900 €',
-    );
-
-    await confirmAndSignAndAssertDone();
-
-    assertDoneScreenFundPension();
-    assertDoneScreenPartialWithdrawal('THIRD');
-  });
-});
-
-describe('withdrawals flow with missing NAV', () => {
-  beforeEach(() => {
-    pensionAccountStatementBackend(server, [
-      {
-        fund: {
-          fundManager: { name: 'Tuleva' },
-          isin: 'INVALID_NAV',
-          name: 'Tuleva World Stocks Pension Fund',
-          managementFeeRate: 0.0034,
-          pillar: 2,
-          ongoingChargesFigure: 0.0039,
-          status: 'ACTIVE',
-          inceptionDate: '2017-01-01',
-          nav: 1,
-        },
-        value: 15000.0,
-        unavailableValue: 0,
-        currency: 'EUR',
-        activeContributions: false,
-        contributions: 12345.67,
-        subtractions: 0,
-        profit: 2654.33,
-      },
-    ]);
-    withdrawalsEligibilityBackend(server);
-  });
-
-  test('can not submit when a NAV is missing', async () => {
-    const partialWithdrawalSizeInput = await screen.findByLabelText(
-      'Do you wish to make a partial withdrawal immediately?',
-      { exact: false },
-    );
-
-    userEvent.type(partialWithdrawalSizeInput, '20000');
-
-    userEvent.click(nextButton());
-
-    await enterIban('EE591254471322749514');
-    userEvent.click(nextButton());
-
-    expect(
-      await screen.findByText(/I submit the following applications and am aware of their terms/i),
-    ).toBeInTheDocument();
-
-    assertMandateCount(4);
-
-    await confirmAndSignAndAssertFailed();
-  });
-});
-
-const nextButton = () => screen.getByRole('button', { name: 'Continue' });
-const confirmationCheckbox = () => screen.getByRole('checkbox');
-const signButton = () => screen.getByRole('button', { name: /Sign/ });
-
-const enterIban = async (iban: string) => {
-  const ibanInput = await screen.findByLabelText('Bank account number (IBAN)');
-  userEvent.type(ibanInput, iban);
-};
-
-const confirmAndSignAndAssertFailed = async () => {
-  userEvent.click(confirmationCheckbox());
-  expect(signButton()).toBeEnabled();
-
-  userEvent.click(signButton());
-
-  expect(
-    screen.getByText(
-      /Submitting withdrawal applications failed. Please contact us: tuleva@tuleva.ee/i,
-    ),
-  ).toBeInTheDocument();
-
-  expect(signButton()).toBeInTheDocument();
-};
-
-const confirmAndSignAndAssertDone = async (singleApplication = false) => {
-  userEvent.click(confirmationCheckbox());
-  expect(signButton()).toBeEnabled();
-
-  userEvent.click(signButton());
-
-  expect(
-    await screen.findByRole(
-      'heading',
-      {
-        name: singleApplication
-          ? 'The application has been submitted'
-          : 'The applications have been submitted',
-      },
-      { timeout: 10_000 },
-    ),
-  ).toBeInTheDocument();
-
-  expect(
-    screen.getByText(
-      singleApplication
-        ? /We have sent the digitally signed application to your email address./i
-        : /We have sent the digitally signed applications to your email address./i,
-    ),
-  ).toBeInTheDocument();
-};
-
-const assertDoneScreenFundPension = () => {
-  expect(
-    screen.getByText(/The first monthly fund pension payment will be sent on/i),
-  ).toBeInTheDocument();
-};
-
-const assertDoneScreenPartialWithdrawal = (pillar: 'SECOND' | 'THIRD') => {
-  let paymentDeadline: HTMLElement;
-
-  if (pillar === 'SECOND') {
-    paymentDeadline = screen.getByText(/The partial withdrawal from II pillar will be sent on/i);
-    expect(within(paymentDeadline).getByText(/16 – 20 January/)).toBeInTheDocument();
-  } else {
-    paymentDeadline = screen.getByText(/The partial withdrawal from III pillar will be sent on/i);
-    expect(within(paymentDeadline).getByText(/four working days/)).toBeInTheDocument();
-  }
-};
-
-const assertDoneScreenSecondPillarWarning = () => {
-  expect(
-    screen.getByText('contributions to your second pillar will stop permanently'),
-  ).toBeInTheDocument();
-};
-
-const assertMandateCount = async (count: number) =>
-  Promise.all(
-    Array(count)
-      .fill(null)
-      .map(async (_, i) =>
-        Promise.all([expect(await screen.findByText(`Application #${i + 1}`)).toBeInTheDocument()]),
-      ),
-  );
-
-const assertFundPensionCalculations = async (
-  fundPensionMonthlySize: string,
-  liquidatedMonthlyPercentage = '0.42%',
-  returnsRegex = /will earn returns for the next 20 years/i,
-) => {
-  const explanationText = screen.getByText(/Every month you will receive/i);
-  expect(within(explanationText).getByText(liquidatedMonthlyPercentage)).toBeInTheDocument();
-
-  expect(screen.getByText(new RegExp(fundPensionMonthlySize, 'i'))).toBeInTheDocument();
-  expect(screen.getByText(returnsRegex)).toBeInTheDocument();
-};
-
-const assertTotalTaxText = (amount: string) => {
-  const taxText = screen.getByText(/Partial withdrawal will be subject to 10% income tax/i);
-  expect(within(taxText).getByText(amount)).toBeInTheDocument();
-};
-
-const assertFundPensionMandate = async (
-  pillar: WithdrawalMandateDetails['pillar'],
-  fundPensionSize: string,
-  bankruptciesPresent?: 'BANKRUPTCIES_ARRESTS_PRESENT',
-) => {
-  const fundPensionSection = await screen.findByTestId(`FUND_PENSION_OPENING_${pillar}`);
-
-  if (pillar === 'SECOND') {
-    expect(
-      await within(fundPensionSection).findByRole('heading', {
-        name: 'Monthly fund pension payments from II pillar',
-      }),
-    ).toBeInTheDocument();
-  } else {
-    expect(
-      await within(fundPensionSection).findByRole('heading', {
-        name: 'Monthly fund pension payments from III pillar',
-      }),
-    ).toBeInTheDocument();
-  }
-
-  const loader = within(fundPensionSection).queryByRole('progressbar');
-  if (loader) {
-    await waitForElementToBeRemoved(loader);
-  }
-
-  expect(
-    within(fundPensionSection).getByText(
-      /The recommended duration is calculated based on the average years left to live according to your age. Currently, it is \d+ years./i,
-    ),
-  ).toBeInTheDocument();
-
-  const paymentDeadline = within(fundPensionSection).getByText(/The first payment will be sent on/);
-
-  expect(within(paymentDeadline).getByText(/16 – 20 January/)).toBeInTheDocument();
-  expect(within(paymentDeadline).getByText(new RegExp(fundPensionSize))).toBeInTheDocument();
-
-  if (pillar === 'SECOND') {
-    const pillarStoppedWarning = within(fundPensionSection).getByText(
-      /Upon submitting the application/,
-    );
-    expect(
-      within(pillarStoppedWarning).getByText(/II pillar contributions will stop/),
-    ).toBeInTheDocument();
-    expect(within(pillarStoppedWarning).getByText(/from August 1/)).toBeInTheDocument();
-
-    if (bankruptciesPresent) {
-      expect(
-        within(fundPensionSection).getByText(
-          /because there is an active arrest or bankruptcy claim registered in the pension registry./,
-        ),
-      ).toBeInTheDocument();
-    }
-  }
-
-  if (!bankruptciesPresent && pillar !== 'SECOND') {
-    const cancelDeadline = within(fundPensionSection).getByText(
-      /I can cancel the application until/,
-    );
-    expect(within(cancelDeadline).getByText(/December 31 at 21:59/)).toBeInTheDocument();
-  }
-};
-
-const assertPartialWithdrawalMandate = async (
-  pillar: WithdrawalMandateDetails['pillar'],
-  rows: { fundName: string; liquidationAmount: string }[],
-  partialWithdrawalAmount: string,
-  bankruptciesPresent?: 'BANKRUPTCIES_ARRESTS_PRESENT',
-) => {
-  const partialWithdrawalSection = await screen.findByTestId(`PARTIAL_WITHDRAWAL_${pillar}`);
-
-  if (pillar === 'SECOND') {
-    expect(
-      await within(partialWithdrawalSection).findByRole('heading', {
-        name: 'Partial withdrawal from II pillar',
-      }),
-    ).toBeInTheDocument();
-  } else {
-    expect(
-      await within(partialWithdrawalSection).findByRole('heading', {
-        name: 'Partial withdrawal from III pillar',
-      }),
-    ).toBeInTheDocument();
-  }
-
-  const loader = within(partialWithdrawalSection).queryByRole('progressbar');
-  if (loader) {
-    await waitForElementToBeRemoved(loader);
-  }
-
-  rows.forEach(({ fundName, liquidationAmount }) => {
-    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-    const rowElement = within(partialWithdrawalSection).getByText(
-      fundName,
-      // eslint-disable-next-line testing-library/no-node-access
-    ).parentElement!;
-
-    expect(within(rowElement).getByText(fundName)).toBeInTheDocument();
-    expect(within(rowElement).getByText(liquidationAmount)).toBeInTheDocument();
-  });
-
-  const paymentDeadline = within(partialWithdrawalSection).getByText(/The payment will be sent on/);
-
-  if (pillar === 'SECOND') {
-    expect(within(paymentDeadline).getByText(/16 – 20 January/)).toBeInTheDocument();
-  } else {
-    expect(within(paymentDeadline).getByText(/in four working days/)).toBeInTheDocument();
-  }
-
-  expect(within(partialWithdrawalSection).getByText(/10% income tax/)).toBeInTheDocument();
-
-  expect(
-    within(partialWithdrawalSection).getByText(new RegExp(partialWithdrawalAmount)),
-  ).toBeInTheDocument();
-
-  if (pillar === 'SECOND') {
-    const pillarStoppedWarning = within(partialWithdrawalSection).getByText(
-      /Upon submitting the application/,
-    );
-    expect(
-      within(pillarStoppedWarning).getByText(/II pillar contributions will stop/),
-    ).toBeInTheDocument();
-    expect(within(pillarStoppedWarning).getByText(/from August 1/)).toBeInTheDocument();
-
-    if (bankruptciesPresent) {
-      expect(
-        within(partialWithdrawalSection).getByText(
-          /because there is an active arrest or bankruptcy claim registered in the pension registry./,
-        ),
-      ).toBeInTheDocument();
-    }
-  }
-};
